+++ src/Resources/contao/languages/en/modules.php
@@ -1,0 +1,3 @@
+<?php
+
+$GLOBALS['TL_LANG']['FMD'][\HeimrichHannot\AmpBundle\FrontendModule\AmpNavigationModule::TYPE] = ['Navigation menu (AMP)', ''];