+++ src/Resources/contao/languages/en/tl_module.php
@@ -1,0 +1,14 @@
+<?php
+
+$lang = &$GLOBALS['TL_LANG']['tl_module'];
+
+/**
+ * Fields
+ */
+$lang['ampRenderSubItemsAsAccordions'][0] = 'Display sub pages as accordion';
+$lang['ampRenderSubItemsAsAccordions'][1] = 'Choose this option to render sub pages as accordions.';
+
+/**
+ * Legends
+ */
+$lang['amp_legend'] = 'Accelerated Mobile Pages (AMP)';